+++ frontend/src/components/RecentUser/RecentUser.css
@@ -1,0 +1,5 @@
+ .recent-user {
+  border: 1px solid green;
+  color: white;
+  background: linear-gradient(#2B2A2A, #313131);
+ }
