+++ frontend/src/components/RecentUser/RecentUser.js
@@ -1,0 +1,12 @@
+import './RecentUser.css';
+
+export default function RecentUser({props, user}) {
+    return (<div className={'recent-user'}>
+        <div>
+            {user.firstName} {user.lastName}
+            {user.email}
+        </div>
+    </div>);
+}
+
+
